email-agent-project/backend/pyproject.toml
[tool.poetry]
name = "email-agent-backend"
version = "0.1.0"
description = "AI-powered email management system using CrewAI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
python = "^3.12.2"

[tool.poetry.dependencies]
python = "^3.12.2"
# Core dependencies
crewai = "0.86.0"
crewai-tools = "^0.17.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"

# LLM Providers
openai = "^1.54.0"
google-generativeai = "^0.8.0"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.0.0"

# Gmail Integration
google-api-python-client = "^2.149.0"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"

# Database
sqlalchemy = "^2.0.35"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"

# Redis
redis = "^5.0.0"
hiredis = "^3.0.0"

# Utilities
httpx = "^0.27.0"
aiofiles = "^24.1.0"
python-multipart = "^0.0.9"
email-validator = "^2.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.2.0"

# WebSocket support
websockets = "^13.1"
python-socketio = "^5.11.0"

# Development tools
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
flake8 = "^7.1.0"
mypy = "^1.11.0"
ipython = "^8.27.0"

[tool.poetry.group.dev.dependencies]
watchdog = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -q --strict-markers"
asyncio_mode = "auto"